fix: Move the class to the appropriate path
